rename failing tests module to fail
