+++ tests/type_check/fail/var_unbounded.pj
